First commit of the users feat implementation
A lot has been done. This commit includes changes to several views, the
new login view, minor changes in controllers, and the addition of a more
formal Model structure under the class directory.
This commit also includes changes relating to the use of absolute paths,
and the use of a lingua franca