Unification and refactoring of static UI
A header.php and footer.php is now used in every view. The header.php
changes the content between authenticated and anonymous users, and marks
the active page. This also mainly resolves #26 .
A minor change was made to const.php, in order to avoid creating
warnings or notices if there is no session, or if it contains no uid.