Various UI changes, and readdition of standard UI in main index
Again, the standarization of the UI of the main index got somehow lost.
Readded here.
Also moved the propietary css to a single file, as agreed.
Some changes to the UI were made, specially in the stats section.
Also, the csv folder, which holds csv files uploaded and downloaded, was
added to gitignor.