Own styles in header, and message div in auth index page
This commit adds a feature in which own style templates can be added to
front pages via the ownStyles[] array, declaring them before the include
of header.php. The script looks for the styles in the css folder.
This was done in order to add a L&F compliant message box to the auth
index page. The latter now displays errors when the username or password
is not empty.