First paginator approach
A paginator should be included in several pages. This implementations is
based in the main stats page. Currently, it shows a generic paginator in
the bottom of the tracked links, and pages the whole list into divisions
of 10.
The "<<" and ">>" buttons are working, but the numbered one are just
currently placeholders.