Wrong error message in short but valid URL
There was a hard, left-over validation that returned false if the length
of the URL was less than eight. The URL is now at least 4 chars, not 8.
The alias generated was new, but the validation function marked the
alias as 'has-error', and the message was about the existance, which was
not the case. Fixed.